requirements.txt
supabase>=2.0.0
python-dotenv>=1.0.0
anthropic>=0.30.0
streamlit>=1.30.0